@@ -13,12 +13,12 @@
     "babel-plugin-transform-runtime": "^6.7.5",
     "babel-preset-es2015": "^6.4.1",
     "babel-preset-react": "^6.5.0",
-    "eslint": "^2.9.0",
+    "eslint": "^3.5.0",
     "eslint-loader": "^1.3.0",
-    "eslint-plugin-react": "^5.0.1",
+    "eslint-plugin-react": "^6.2.2",
     "express": "^4.13.4",
     "react-hot-loader": "^1.3.0",
-    "webpack": "^1.13.0",
+    "webpack": "^1.13.2",
     "webpack-dev-middleware": "^1.6.1",
     "webpack-hot-middleware": "^2.10.0"
   },
