@@ -1,21 +1,33 @@
 import React, { Component } from 'react'
 import Post from '../components/Post.jsx'
 import PostForm from '../components/PostForm.jsx'
+import 'whatwg-fetch'
 
-let posts = [
- { id: 1, name: 'name 1', description: 'description11' },
- { id: 2, name: 'name 2', description: 'description2' },
- { id: 3, name: 'name 3', description: 'description3' }
-]
 
 export default class App extends Component {
+  componentWillMount() {
+    this.getData()
+  }
+
+  getData() {
+    fetch('http://localhost:3000/api/posts')
+      .then(response => response.json())
+      .then(response => {
+        this.setState({posts: response.result})
+      })
+  }
+
+  handlePostSubmit() {
+    this.getData()
+  }
+
   render() {
     return(
       <div>
-        {posts.map((post) => {
-          return <Post key={post.id} id={post.id} name={post.name} description={post.description} />
+        {this.state && this.state.posts.map((post) => {
+          return <Post key={post.id} id={post.id} username={post.username} title={post.title} body={post.body} />
         })}
-        <PostForm />
+        <PostForm onPostSubmit={this.handlePostSubmit.bind(this)} />
       </div>
     )
   }
